requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple
--extra-index-url https://pypi.org/simple

aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.11.7
    # via fsspec
aiosignal==1.3.1
    # via aiohttp
async-timeout==5.0.1
    # via aiohttp
attrs==24.2.0
    # via aiohttp
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via -r requirements.in
coloredlogs==15.0.1
    # via onnxruntime
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
flatbuffers==24.3.25
    # via onnxruntime
fonttools==4.55.0
    # via matplotlib
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.10.0
    # via
    #   huggingface-hub
    #   pytorch-lightning
    #   torch
huggingface-hub==0.26.2
    # via -r requirements.in
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   requests
    #   yarl
importlib-resources==6.4.5
    # via matplotlib
jinja2==3.1.4
    # via torch
kiwisolver==1.4.7
    # via matplotlib
lightning-utilities==0.11.9
    # via
    #   pytorch-lightning
    #   torchmetrics
markupsafe==3.0.2
    # via jinja2
matplotlib==3.9.2
    # via
    #   -r requirements.in
    #   seaborn
    #   ultralytics
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
networkx==3.2.1
    # via torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   onnx
    #   onnxruntime
    #   opencv-python
    #   pandas
    #   rapidocr-onnxruntime
    #   scipy
    #   seaborn
    #   shapely
    #   torchmetrics
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
onnx==1.17.0
    # via -r requirements.in
onnxruntime==1.19.2
    # via
    #   -r requirements.in
    #   rapidocr-onnxruntime
opencv-python==4.10.0.84
    # via
    #   -r requirements.in
    #   rapidocr-onnxruntime
    #   ultralytics
packaging==24.2
    # via
    #   huggingface-hub
    #   lightning-utilities
    #   matplotlib
    #   onnxruntime
    #   pytorch-lightning
    #   torchmetrics
pandas==2.2.3
    # via
    #   -r requirements.in
    #   seaborn
    #   ultralytics
pillow==11.0.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   rapidocr-onnxruntime
    #   torchvision
    #   ultralytics
propcache==0.2.0
    # via
    #   aiohttp
    #   yarl
protobuf==5.28.3
    # via
    #   onnx
    #   onnxruntime
psutil==6.1.0
    # via ultralytics
py-cpuinfo==9.0.0
    # via ultralytics
pyclipper==1.3.0.post6
    # via
    #   -r requirements.in
    #   rapidocr-onnxruntime
pyparsing==3.2.0
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytorch-lightning==2.4.0
    # via -r requirements.in
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   pytorch-lightning
    #   rapidocr-onnxruntime
    #   ultralytics
rapidocr-onnxruntime==1.3.25
    # via -r requirements.in
requests==2.32.3
    # via
    #   huggingface-hub
    #   ultralytics
scipy==1.13.1
    # via
    #   -r requirements.in
    #   ultralytics
seaborn==0.13.2
    # via
    #   -r requirements.in
    #   ultralytics
shapely==2.0.6
    # via
    #   -r requirements.in
    #   rapidocr-onnxruntime
six==1.16.0
    # via
    #   python-dateutil
    #   rapidocr-onnxruntime
sympy==1.13.1
    # via
    #   onnxruntime
    #   torch
torch==2.5.1
    # via
    #   -r requirements.in
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchmetrics==1.6.0
    # via pytorch-lightning
torchvision==0.20.1
    # via
    #   -r requirements.in
    #   ultralytics
tqdm==4.67.0
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   pytorch-lightning
    #   rapidocr-onnxruntime
    #   ultralytics
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   lightning-utilities
    #   multidict
    #   pytorch-lightning
    #   torch
tzdata==2024.2
    # via pandas
ultralytics==8.3.36
    # via -r requirements.in
ultralytics-thop==2.0.12
    # via ultralytics
unidecode==1.3.8
    # via -r requirements.in
urllib3==2.2.3
    # via requests
yarl==1.18.0
    # via aiohttp
zipp==3.21.0
    # via importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
